pbench-metadata-log changes to record iterations
The benchmark scripts record their iterations in a temp file.
pbench-metadata-log uses that to add an iterations entry to
the metadata log.

pbench-add-metalog-option is a generic script that is used for
that purpose, but it can add an arbitrary option to an arbitrary
section of the metadata log.

Modify unit tests and add a new one to test this functionality.